+++ alinkjs/packages/assetlink-default-plugins/plugins/AssetGroupMembersPageSlotProvider.alink.vue
@@ -102,7 +102,7 @@ const resolveCurrentGroupMembers = async () => {
 
           const currentMember = entitySourceCache.query((q) => q.findRecord({ type: possibleMember.type, id: possibleMember.id }));
 
-          if (currentMember.attributes.status === 'archived') {
+          if (!currentMember || currentMember.attributes.status === 'archived') {
             return [];
           }