+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Add the ability to navigate by clicking on group members and location occupants on the asset page
 - Add a entity searcher plugin to find entities by their old farmOS 1.x urls
 - Accept pasted images in PhotoInput component
+- Expose the AssetLink.fetch method to help support non-JSON:API or certain advanced use-cases
 
 ### Fixed